commit after long int declaration

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-long long number = 612852475143;
+long int number = 612852475143;
 int oddPrimeNum = 3;
 int maxPrimeNum = 0;
